Moved articles loader to another module
