+++ index.html
@@ -16,10 +16,11 @@
       <section class="section-block">
         <h2>News Channel:</h2>
         <select class="channels" name="" id="news-channels"></select>
+        <button type="button" id="show-articles" class="show-articles-btn">Load Articles</button>
       </section>
-      <section class="section-block">
+      <section class="section-block is-hidden js-articles-section">
         <h2>Articles:</h2>
-        <section class="articles" id="news-results"></section>
+        <section class="articles" id="articles-results"></section>
       </section>
     </div>
   </main>
