+++ src/scripts/helper.js
@@ -1,0 +1,12 @@
+// method definitions
+export let dataHelper = {
+  handleErrors(response) {
+    if (!response.ok) {
+      Promise.reject({
+        status : response.status,
+        statusText : response.statusText
+      });
+    }
+    return response;
+  }
+};
